feat: move RequestEmailVerification component to the correct directory and update import path
